Commit 0

+++ src/main/java/io/company/PokemonService.java
@@ -16,19 +16,6 @@ public class PokemonService {
     @Autowired
     Repository repository;
 
-    static List<Pokemon> pokemons = new ArrayList<Pokemon>();
-    static {
-        Pokemon pokemon1 = new Pokemon("paradise", "AmericanExpress",1, 300, 500, "Boing3412");
-        Pokemon pokemon2 = new Pokemon("hellfire", "AmericanExpress",2,  85, 7, "charger");
-        Pokemon pokemon3 = new Pokemon("ribombee", "AmericanExpress",3,   146, 23, "combat");
-
-
-        pokemons.add(pokemon1);
-        pokemons.add(pokemon2);
-        pokemons.add(pokemon3);
-
-
-    }
     public Iterable<Pokemon> getAllPokemons() {
 
         return repository.findAll();
@@ -37,7 +24,14 @@ public class PokemonService {
     public void fakerPokemon() {
 
         Random random = new Random();
+        Pokemon pokemon1 = new Pokemon("paradise", "AmericanExpress",1, 300, 500, "Boing3412");
+        Pokemon pokemon2 = new Pokemon("hellfire", "AmericanExpress",2,  85, 7, "charger");
+        Pokemon pokemon3 = new Pokemon("ribombee", "AmericanExpress",3,   146, 23, "combat");
 
+
+        createPokemon(pokemon1);
+        createPokemon(pokemon2);
+        createPokemon(pokemon3);
         Faker faker = new Faker();
         int number;
         String use = null;
@@ -57,15 +51,15 @@ public class PokemonService {
                     break;
                 }
             }
-            pokemons.add(new Pokemon (faker.pokemon().name(), faker.pokemon().location(), faker.number().numberBetween(1, 856),
+            createPokemon(new Pokemon (faker.pokemon().name(), faker.pokemon().location(), faker.number().numberBetween(1, 856),
                     faker.number().numberBetween(5, 1200),faker.number().numberBetween(2, 12000),
                     use));
         }
 
     }
-    public Pokemon createPokemon (Pokemon pokemon){
+    public void createPokemon (Pokemon pokemon){
 
-        return repository.save(pokemon);
+        repository.save(pokemon);
     }
 
     public Optional<Pokemon> findPokemonById(String id){
@@ -73,34 +67,27 @@ public class PokemonService {
         return repository.findById(id);
     }
 
-    public Optional<Pokemon> findPokemonByNumber(String title){
-        return repository.findPokemonByNumber(title);
+    public Optional<Pokemon> findPokemonByNumber(int number){
+        return repository.findPokemonByNumber(number);
     }
 
-    public Pokemon deletePokemonByNumber(int number){
-        //Find out IF this id-book IS in our DB
-        Optional<Pokemon> deletedPokemon = repository.deleteByNumber(number);
-        //
-        return null;
+    public void deletePokemonByNumber(int number){
+        if(findPokemonByNumber(number).isPresent()){
+            Optional<Pokemon> deletedPokemon = repository.deleteByNumber(number);
+        }
     }
 
     public void deletePokemonById(String id){
-        repository.deleteById(id);
+        if(findPokemonById(id).isPresent()){
+//            Optional<Pokemon> deletedPokemon = repository.deleteById(id);
+        }
     }
 
-    public Pokemon updatePokemon (Pokemon book){
-        return repository.save(book);
+    public Pokemon updatePokemon (Pokemon pokemon){
+
+
+
+        return repository.save(pokemon);
     }
-
-
-
-    public List<Pokemon> queryPokemonsFromArray() {
-
-        System.out.println("Pokemons" + pokemons);
-
-        return pokemons;
-    }
-
-
 
 }
